debug

+++ src/admin/form_handlers/synonym.py
@@ -1,6 +1,6 @@
 from flask import request, flash, url_for, render_template
 from flask_classy import FlaskView, route
-from flask_login import login_required
+from flask_login import login_required, current_user
 from markupsafe import Markup
 from sqlalchemy import exc
 from werkzeug.utils import redirect
@@ -15,6 +15,8 @@ class FormHandlerSynonym(FlaskView):
     @route('/add_synonym', methods=["POST"])
     @login_required
     def add_synonym(self):
+        if not current_user.right_synonym:
+            return render_template('admin/access_denied.html')
         add_synonyms_dependent_form = AddSynonymsDependentForm()
         main_word_id = request.args.get("word_id")
         if not add_synonyms_dependent_form.validate_on_submit():
@@ -23,15 +25,21 @@ class FormHandlerSynonym(FlaskView):
             word = add_synonyms_dependent_form.word.data.lower()
             word = " ".join(word.split())
 
-            if not word:
-                flash('Неправильно заполнены поля', category='danger')
+            try:
+                synonyms = SynonymousWords.query
+                word_from_db = synonyms.filter(SynonymousWords.word == word).first()
+            except (NameError, AttributeError):
+                return render_template("admin/error_page.html", message="Ошибка чтения из БД")
+            if word_from_db is not None:
+                flash('Такое слово уже существует', category='danger')
             else:
                 if main_word_id:
-                    main_word = SynonymousWords.query.get(main_word_id)
+                    main_word = synonyms.get(main_word_id)
                     if main_word.synonym_id:
                         flash('Главный синоним не должен сам быть чьим-то синонимом. Фильтры сброшены',
                               category='danger')
-                        return redirect(url_for('.ViewTechSetting:synonyms'))
+                        return redirect(url_for('.ViewSynonym:synonyms'))
+
                     synonym = SynonymousWords(word=word, synonym_id=main_word_id)
                 else:
                     synonym = SynonymousWords(word=word)
@@ -42,12 +50,14 @@ class FormHandlerSynonym(FlaskView):
                 except exc.SQLAlchemyError:
                     flash('Ошибка внесения изменений в базу данных', category='danger')
         if main_word_id:
-            return redirect(url_for('.ViewTechSetting:synonyms_replacement', word_id=main_word_id))
-        return redirect(url_for('.ViewTechSetting:synonyms'))
+            return redirect(url_for('.ViewSynonym:synonyms_replacement', word_id=main_word_id))
+        return redirect(url_for('.ViewSynonym:synonyms'))
 
     @route('/edit_synonym', methods=["POST"])
     @login_required
     def edit_synonym(self):
+        if not current_user.right_synonym:
+            return render_template('admin/access_denied.html')
         edit_synonyms_dependent_form = EditSynonymsDependentForm()
         main_word_id = request.args.get("word_id")
         if not edit_synonyms_dependent_form.validate_on_submit():
@@ -56,16 +66,19 @@ class FormHandlerSynonym(FlaskView):
             word_id = edit_synonyms_dependent_form.word_id.data
             word = edit_synonyms_dependent_form.word.data.lower()
             word = " ".join(word.split())
-            if not word:
-                flash('Неправильно заполнены поля', category='danger')
-            else:
-                try:
-                    synonym = SynonymousWords.query.get(word_id)
-                except NameError:
-                    return render_template("admin/error_page.html", message="Ошибка чтения из БД")
 
-                if synonym is None:
-                    flash('Нет такого слова', category='danger')
+            try:
+                synonyms = SynonymousWords.query
+                word_from_db = synonyms.filter(SynonymousWords.word == word).first()
+                synonym = SynonymousWords.query.get(word_id)
+            except NameError:
+                return render_template("admin/error_page.html", message="Ошибка чтения из БД")
+
+            if synonym is None:
+                flash('Нет такого слова', category='danger')
+            else:
+                if word_from_db is not None and word_from_db != synonym:
+                    flash('Такое слово уже существует. Изменения не сохранены', category='danger')
                 else:
                     synonym.word = word
                     try:
@@ -74,12 +87,14 @@ class FormHandlerSynonym(FlaskView):
                     except exc.SQLAlchemyError:
                         flash('Ошибка внесения изменений в базу данных', category='danger')
         if main_word_id:
-            return redirect(url_for('.ViewTechSetting:synonyms_replacement', word_id=main_word_id))
-        return redirect(url_for('.ViewTechSetting:synonyms'))
+            return redirect(url_for('.ViewSynonym:synonyms_replacement', word_id=main_word_id))
+        return redirect(url_for('.ViewSynonym:synonyms'))
 
     @route('/delete_synonym', methods=["POST"])
     @login_required
     def delete_synonym(self):
+        if not current_user.right_synonym:
+            return render_template('admin/access_denied.html')
         delete_synonyms_dependent_form = DeleteSynonymsDependentForm()
         main_word_id = request.args.get("word_id")
         if not delete_synonyms_dependent_form.validate_on_submit():
@@ -116,5 +131,5 @@ class FormHandlerSynonym(FlaskView):
                     except exc.SQLAlchemyError:
                         flash('Ошибка внесения изменений в базу данных', category='danger')
         if main_word_id:
-            return redirect(url_for('.ViewTechSetting:synonyms_replacement', word_id=main_word_id))
-        return redirect(url_for('.ViewTechSetting:synonyms'))
+            return redirect(url_for('.ViewSynonym:synonyms_replacement', word_id=main_word_id))
+        return redirect(url_for('.ViewSynonym:synonyms'))
